third_party/WebKit/LayoutTests/fast/mediastream/MediaStreamTrack-getSourceInfos-expected.txt

Tests MediaStreamTrack::getSourceInfos.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".


PASS sources = MediaStreamTrack.getSourceInfos(); did not throw exception.
PASS sources.length > 0 is true
PASS sources = MediaStreamTrack.getSourceInfos(); did not throw exception.
PASS sources.length > 0 is true
PASS sources[0].id === previousId is true
PASS sources[0].label.length > 0 is true
PASS successfullyParsed is true

TEST COMPLETE
